moving all pages under orgProfile
