+++ pages/[organisationName]/orgProfile.tsx
@@ -9,8 +9,8 @@ import {
   } from "@material-ui/core"
   import { makeStyles } from "@material-ui/core/styles"
   import React from "react"
-  import { BasePage } from "../components/templates"
-import { Footer } from "../components/templates/Footer";
+  import { BasePage } from "../../components/templates"
+import { Footer } from "../../components/templates/Footer";
 import {DropzoneArea} from 'material-ui-dropzone'
 
   const useStyles = makeStyles((theme) => ({
