+++ pages/[organisationName]/paymentForm.tsx
@@ -6,9 +6,9 @@ import Step from '@material-ui/core/Step';
 import StepLabel from '@material-ui/core/StepLabel';
 import Button from '@material-ui/core/Button';
 import Typography from '@material-ui/core/Typography';
-import { BasePage } from '../components/templates';
+import { BasePage } from '../../components/templates';
 import { Container, NoSsr } from '@material-ui/core';
-import {DonationContributorStep, PaymentCardStep, PaymentSummaryStep, RecurringPaymentStep}  from '../components/templates/donationSteps';
+import {DonationContributorStep, PaymentCardStep, PaymentSummaryStep, RecurringPaymentStep}  from '../../components/templates/donationSteps';
 
 
 const useStyles = makeStyles((theme) => ({
