+++ tsconfig.json
@@ -23,5 +23,5 @@
     "isolatedModules": true
   },
   "exclude": ["node_modules"],
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/temp.js"]
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "pages/temp.js", "pages/[organisationName]"]
 }
